requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
prophet>=1.1.0
pyyaml>=6.0
cassandra-driver>=3.28.0
pyspark>=3.4.0
matplotlib>=3.7.0
statsmodels>=0.14.0
